TDT4225 - Store, distribuerte datamengder/assignment2/requirements.txt
haversine==2.7.0
mysql-connector-python==8.0.30
tabulate==0.8.9
